Solution to Check Alphabet Case

+++ Check Alphabet Case@@6788dd63490637cfc44e1b92/code.c
@@ -3,7 +3,7 @@ int main()
 {
     char a;
     scanf("%c", &a);
-    if (65<=a<=90)
+    if (65<=(int)a<=90)
     {
         printf("UpperCase");
     }
